\name{as.XML-methods}
\docType{methods}
\alias{as.XML,SDMX-method}
\alias{as.XML}
\title{as.XML}
\description{a convenience method to get SDMX as XML object}

\usage{
as.XML(obj)
}

\arguments{
	\item{obj}{ An object deriving from class "SDMX"}
}

\section{Methods}{
\describe{
	\item{obj = "SDMX"}{ An object deriving from class "SDMX"}
}
}

\value{a object of class "XMLInternalDocument"}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{methods}
\keyword{XML}
