\name{codelists.SDMXCodelists}
\Rdversion{1.1}
\alias{codelists.SDMXCodelists}
\title{codelists.SDMXCodelists}

\description{function used to parse the list of codelists of a SDMX-ML document}

\usage{
  codelists.SDMXCodelists(xmlObj)
}

\arguments{
  \item{xmlObj}{an object of class "XMLInternalDocument"}
}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\section{Warning }{\code{codelists.SDMXCodelists} is a function used internally to parse the list of codelists during instantiation of "SDMXCodelists" objects}

\seealso{\link{SDMXCodelists-class} \link{SDMXCodelist-class}}