\name{conceptSchemes.SDMXConcepts}
\Rdversion{1.1}
\alias{conceptSchemes.SDMXConcepts}
\title{conceptSchemes.SDMXConcepts}

\description{function used to parse the list of concepts Schemes of a SDMX-ML documents, as defined from SDMX 2.0 specifications}

\usage{
  conceptSchemes.SDMXConcepts(xmlObj)
}

\arguments{
  \item{xmlObj}{an object of class "XMLInternalDocument"}
}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\section{Warning }{\code{concepts.SDMXConcepts} is a function used internally to parse the list of concepts during instantiation of "SDMXConcepts" objects}

\seealso{\link{SDMXConcepts-class} \link{SDMXConceptScheme-class} \link{SDMXConcept-class}}