\name{datastructures.SDMXDataStructures}
\Rdversion{1.1}
\alias{datastructures.SDMXDataStructures}
\title{datastructures.SDMXDataStructures}

\description{function used to parse the list of datastructures / keyfamilies of a SDMX-ML document}

\usage{
  datastructures.SDMXDataStructures(xmlObj)
}

\arguments{
  \item{xmlObj}{an object of class "XMLInternalDocument"}
}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\section{Warning }{\code{datastructures.SDMXDataStructures} is a function used internally to parse the list of datastructure / keyfamily during instantiation of "SDMXDataStructures" objects}

\seealso{\link{SDMXDataStructures-class} \link{SDMXDataStructure-class}}