\name{getSDMXFooter-methods}
\docType{methods}
\alias{getSDMXFooter,SDMX-method}
\alias{getSDMXFooter}
\title{getSDMXFooter}
\description{Access the SDMX Footer}

\usage{
getSDMXFooter(obj)
}

\arguments{
  \item{obj}{An object deriving from class "SDMX"}
}

\section{Methods}{
\describe{
	\item{obj = "SDMX"}{An object deriving from class "SDMX"}
}}

\value{an object of class "SDMXFooter"}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{methods}

\seealso{\link{SDMXFooter-class}}