\name{getSDMXHeader-methods}
\docType{methods}
\alias{getSDMXHeader,SDMX-method}
\alias{getSDMXHeader}
\title{getSDMXHeader}
\description{Access the SDMX header}

\usage{
getSDMXHeader(obj)
}

\arguments{
	\item{obj}{An object deriving from class "SDMX"}
}

\section{Methods}{
\describe{
	\item{obj = "SDMX"}{An object deriving from class "SDMX"}
}}

\value{an object of class "SDMXHeader"}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{methods}

\seealso{\link{SDMXHeader-class}}