\name{getSDMXSchema-methods}
\docType{methods}
\alias{getSDMXSchema,SDMX-method}
\alias{getSDMXSchema}
\title{getSDMXSchema}
\description{Access the SDMX-ML schema version}

\usage{
getSDMXSchema(obj)
}

\arguments{
	\item{obj}{ An object deriving from class "SDMX"}
}

\section{Methods}{
\describe{
	\item{obj = "SDMX"}{An object deriving from class "SDMX"}
}}

\value{an object of class "SDMXSchema"}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{methods}

\seealso{\link{SDMXSchema-class}}
