\name{getSDMXType-methods}
\docType{methods}
\alias{getSDMXType,SDMX-method}
\alias{getSDMXType}
\title{getSDMXType}
\description{Access the type of SDMX object}

\usage{
getSDMXType(obj)
}

\arguments{
	\item{obj}{An object deriving from class "SDMX"}
}

\section{Methods}{
\describe{
	\item{obj}{An object deriving from class "SDMX"}
}}

\value{an object of class "SDMXType"}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{methods}

\seealso{\link{SDMXType-class}}
