\name{getStructureType-methods}
\docType{methods}
\alias{getStructureType,SDMXStructureType-method}
\alias{getStructureType}
\title{getStructureType}
\description{Access the subtype of SDMXStructureType object}

\usage{
getStructureType(obj)
}

\arguments{
	\item{obj}{An object deriving from class "SDMXStructureType"}
}

\section{Methods}{
\describe{
	\item{obj}{An object deriving from class "SDMXStructureType"}
}}

\value{an object of class "character"}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{methods}

\seealso{\link{SDMXStructureType-class}}
