\name{getType-methods}
\docType{methods}
\alias{getType,SDMXType-method}
\alias{getType}
\title{getType}
\description{Access the type of SDMXType object}

\usage{
getType(obj)
}

\arguments{
	\item{obj}{An object deriving from class "SDMXType"}
}

\section{Methods}{
\describe{
	\item{obj}{An object deriving from class "SDMXType"}
}}

\value{an object of class "character"}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{methods}

\seealso{\link{SDMXType-class} \link{getSDMXType}}
