\name{getVersion-methods}
\docType{methods}
\alias{getVersion,SDMXSchema-method}
\alias{getVersion}
\title{getVersion}
\description{Access the version of a SDMXSchema object}

\usage{
getVersion(obj)
}

\arguments{
	\item{obj}{An object deriving from class "SDMXSchema"}
}

\section{Methods}{
\describe{
	\item{obj}{An object deriving from class "SDMXSchema"}
}}

\value{an object of class "character"}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{methods}

\seealso{\link{SDMXSchema-class} \link{getSDMXSchema}}
