% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMX-methods.R
\name{isRegistryInterfaceEnvelope}
\alias{isRegistryInterfaceEnvelope}
\title{isRegistryInterfaceEnvelope}
\usage{
isRegistryInterfaceEnvelope(xmlObj, nativeRoot)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{nativeRoot}{object of class "logical" indicating if it is the native document}
}
\value{
an object of class "logical"
}
\description{
function used to detect if the XML document corresponds to a 
             registry interface query
}
\section{Warning}{

\code{isRegistryInterfaceEnvelope} is a function used internally by \link{readSDMX}
}

\seealso{
\link{SDMX-class} \link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
