\name{primarymeasure.SDMXComponents}
\Rdversion{1.1}
\alias{primarymeasure.SDMXComponents}
\title{primarymeasure.SDMXComponents}

\description{function used to parse the primarymeasure of a SDMX-ML document}

\usage{
  primarymeasure.SDMXComponents(xmlObj)
}

\arguments{
  \item{xmlObj}{an object of class "XMLInternalDocument"}
}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\section{Warning }{\code{primarymeasure.SDMXComponents} is a function used internally to parse the primarymeasure during instantiation of "SDMXComponents" objects (used in DataStructures / KeyFamilies)}

\seealso{\link{SDMXComponents-class} \link{SDMXDataStructure-class}}