\name{rsdmx}
\Rdversion{1.1}
\alias{rsdmx-package}
\alias{rsdmx}


\docType{package}

\title{A R package to read SDMX data}

\description{RSDMX is a package to parse/read SDMX data in R. The package is for now at its earlier stages 
and the aim is to make it evolving in the future. It will provide classes and methods to read SDMX data.}


\details{
\tabular{ll}{
Package: \tab rsdmx\cr
Type: \tab Package\cr
Version
: \tab 0.2\cr
Date: \tab 2014-04-22\cr
License: \tab GPL(>=2.0)\cr
LazyLoad: \tab yes\cr
}

RSDMX is a package to parse/read SDMX data in R. The package is for now at its earlier stages 
and the aim is to make it evolving in the future. It will provide classes and methods to read SDMX data.

}

\author{Emmanuel Blondel <emmanuel.blondel1@gmail.com>}


\keyword{SDMX}
\keyword{statistics}
\keyword{timeseries}
