% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMXStructureType-methods.R, R/SDMX-methods.R,
%   R/SDMXData-methods.R
\docType{methods}
\name{getStructureType}
\alias{getStructureType}
\alias{getStructureType,SDMXStructureType-method}
\alias{getStructureType,SDMXStructureType,ANY-method}
\alias{getNamespaces}
\alias{getNamespaces,SDMX-method}
\alias{getNamespaces,SDMX,ANY-method}
\alias{setDSD,SDMXData,SDMXDataStructureDefinition-method}
\title{getNamespaces}
\usage{
getStructureType(obj)

\S4method{getStructureType}{SDMXStructureType}(obj)

getNamespaces(obj)

\S4method{getNamespaces}{SDMX}(obj)

\S4method{setDSD}{SDMXData,SDMXDataStructureDefinition}(obj, dsd)
}
\arguments{
\item{obj}{An object deriving from class "SDMX"}
}
\value{
an object of class \code{data.frame} giving the id and uri for each 
        of the namespaces handled in the SDMX-ML document.
}
\description{
Access the namespaces of the SDMX-ML object
}
\seealso{
\link{SDMX-class}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
