% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-SDMXAgency.R, R/SDMXAgency-methods.R
\docType{class}
\name{SDMXAgency}
\alias{SDMXAgency}
\alias{SDMXAgency,SDMXAgency-method}
\alias{SDMXAgency-class}
\title{Class "SDMXAgency"}
\usage{
SDMXAgency(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXAgency"
}
\description{
A basic class to handle a SDMX Concept
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
\seealso{
\link{readSDMX}
}

