% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-SDMXCodelists.R, R/SDMXCodelists-methods.R
\docType{class}
\name{SDMXCodelists}
\alias{SDMXCodelists}
\alias{SDMXCodelists,SDMXCodelists-method}
\alias{SDMXCodelists-class}
\title{Class "SDMXCodelists"}
\usage{
SDMXCodelists(xmlObj)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}
}
\value{
an object of class "SDMXCodelists"
}
\description{
A basic class to handle SDMX Codelists
}
\section{Slots}{

\describe{
\item{\code{codelists}}{Object of class "list" giving the list of "SDMXCodelist"}
}}
\section{Warning}{

This class is not useful in itself, but \link{SDMXDataStructureDefinition} objects
will encapsulate it as slot, when parsing an SDMX-ML document.
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
\seealso{
\link{readSDMX}
}

