\name{SDMXCompactData-class}
\docType{class}
\alias{SDMXCompactData-class}
\alias{SDMXCompactData}
\alias{SDMXCompactData-method}


\title{Class "SDMXCompactData"}
\description{ A basic class to handle a SDMX-ML compact data set}
\section{Objects from the Class}{are never to be generated; will be used by the \link{readSDMX} main parser function}


\usage{
SDMXCompactData(xmlObj)
}

\arguments{
\item{xmlObj}{an object of class "XMLInternalDocument"}
}

\author{ Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{classes}