% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXConcepts.R, R/SDMXConcepts-methods.R
\docType{class}
\name{SDMXConcepts}
\alias{SDMXConcepts}
\alias{SDMXConcepts-class}
\alias{SDMXConcepts,SDMXConcepts-method}
\title{Class "SDMXConcepts"}
\usage{
SDMXConcepts(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXConcepts"
}
\description{
A basic class to handle SDMX Concepts
}
\section{Slots}{

\describe{
\item{\code{concepts}}{Object of class "list" giving the list of "SDMXConcept". This 
slot is available to ensure backward compatibility with SDMX 1.0 in SDMX 
2.0 or 2.1 documents}

\item{\code{conceptSchemes}}{Object of class "list" giving the list of "SDMXConceptScheme", 
which will encapsulate the list of "SDMXConcept" (defined from SDMX 2.0)}
}}

\section{Warning}{

This class is not useful in itself, but all SDMX non-abstract classes will 
encapsulate it as slot, when parsing an SDMX-ML document (Concepts, or 
DataStructureDefinition)
}

\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
