% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXCrossSectionalData.R,
%   R/SDMXCrossSectionalData-methods.R
\docType{class}
\name{SDMXCrossSectionalData}
\alias{SDMXCrossSectionalData}
\alias{SDMXCrossSectionalData-class}
\alias{SDMXCrossSectionalData,SDMXCrossSectionalData-method}
\title{Class "SDMXCrossSectionalData"}
\usage{
SDMXCrossSectionalData(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXCrossSectionalData"
}
\description{
A basic class to handle a SDMX-ML cross sectional data set
}
\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
