% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMXData-methods.R
\docType{methods}
\name{setDSD}
\alias{setDSD}
\alias{setDSD,SDMXData,SDMXDataStructureDefinition-method}
\title{setDSD}
\usage{
setDSD(obj, dsd)
}
\arguments{
\item{obj}{An object deriving from class "SDMXData"}

\item{dsd}{An object of class "SDMXDataStructureDefinition"}
}
\value{
the 'obj' object of class "SDMXData" enriched with the dsd
}
\description{
set the dsd slot of a \code{SDMXData} object
}
\seealso{
\link{SDMXData-class}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
