% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXDataFlows.R,
%   R/SDMXDataFlows-methods.R
\docType{class}
\name{SDMXDataFlows}
\alias{SDMXDataFlows}
\alias{SDMXDataFlows-class}
\alias{SDMXDataFlows,SDMXDataFlows-method}
\title{Class "SDMXDataFlows"}
\usage{
SDMXDataFlows(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXDataFlows"
}
\description{
A basic class to handle a SDMX DataFlows
}
\section{Slots}{

\describe{
\item{\code{dataflows}}{Object of class "list" giving the list of DataFlows (datasets), 
(see \link{SDMXDataFlow})}
}}

\section{Warning}{

This class is not useful in itself, but all SDMX non-abstract classes will 
encapsulate it as slot, when parsing an SDMX-ML document (Concepts, or 
DataStructureDefinition)
}

\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
