\name{SDMXDataSet-class}
\docType{class}
\alias{SDMXDataSet-class}
\alias{SDMXDataSet}
\alias{SDMXDataSet-method}


\title{Class "SDMXDataSet"}
\description{ A basic class to handle a SDMX-ML dataset}
\section{Objects from the Class}{are never to be generated; will be used by the \link{readSDMX} main parser function}


\usage{
SDMXDataSet(xmlObj)
}

\arguments{
\item{xmlObj}{an object of class "XMLInternalDocument"}
}

\author{ Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{classes}