% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXDataStructureDefinition.R,
%   R/SDMXDataStructureDefinition-methods.R
\docType{class}
\name{SDMXDataStructureDefinition}
\alias{SDMXDataStructureDefinition}
\alias{SDMXDataStructureDefinition-class}
\alias{SDMXDataStructureDefinition,SDMXDataStructureDefinition-method}
\title{Class "SDMXDataStructureDefinition"}
\usage{
SDMXDataStructureDefinition(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXDataStructureDefinition"
}
\description{
A basic class to handle a SDMX DataStructureDefinition (DSD)
}
\section{Slots}{

\describe{
\item{\code{organisationSchemes}}{Object of class "SDMXOrganisationSchemes" giving the
list of agencies (see \link{SDMXAgencyScheme} and \link{SDMXAgency})}

\item{\code{concepts}}{Object of class "SDMXConcepts" giving the list of concepts or 
conceptSchemes (see \link{SDMXConcepts})}

\item{\code{codelists}}{Object of class "SDMXCodelists" giving the list of codelists 
(see \link{SDMXCodelists})}

\item{\code{datastructures}}{Object of class "SDMXDataStructures" giving the list of 
datastructures /key families (see \link{SDMXDataStructures})}
}}

\section{Warning}{

This class is not useful in itself, but all SDMX non-abstract classes will 
encapsulate it as slot, when parsing an SDMX-ML document (Concepts, or 
DataStructureDefinition)
}

\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
