% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-SDMXDotStatRequestBuilder.R, R/SDMXDotStatRequestBuilder-methods.R
\docType{class}
\name{SDMXDotStatRequestBuilder}
\alias{SDMXDotStatRequestBuilder}
\alias{SDMXDotStatRequestBuilder,SDMXDotStatRequestBuilder-method}
\alias{SDMXDotStatRequestBuilder-class}
\title{Class "SDMXDotStatRequestBuilder"}
\usage{
SDMXDotStatRequestBuilder(regUrl, repoUrl, unsupportedResources,
                           skipProviderId, forceProviderId)
}
\arguments{
\item{regUrl}{an object of class "character" giving the base Url of the SDMX
service registry}

\item{repoUrl}{an object of class "character" giving the base Url of the SDMX
service repository}

\item{unsupportedResources}{an object of class "list" giving eventual unsupported
REST resources. Default is an empty list object}

\item{skipProviderId}{an object of class "logical" indicating that the provider
agencyIdshould be skipped. Used to control lack of strong SDMX REST compliance
from data providers. For now, it applies only for the "data" resource.}

\item{forceProviderId}{an object of class "logical" indicating if the provider
       agencyId has to be added at the end of the request. Default value is
       \code{FALSE}. For some providers, the \code{all} value for the provider
       agency id is not allowed, in this case, the \code{agencyId} of the data
       provider has to be forced in the web-request}
}
\description{
A experimental class to handle a SDMX DotStat (*.Stat) service
             request builder
}
\section{Slots}{

\describe{
\item{\code{regUrl}}{an object of class "character" giving the base Url of the SDMX service registry}

\item{\code{repoUrl}}{an object of class "character" giving the base Url of the SDMX service repository}

\item{\code{compliant}}{an object of class "logical" indicating if the request builder is somehow compliant with a service specification}
}}
\section{Warning}{

This class is not useful in itself, but all SDMX non-abstract classes will
encapsulate it as slot, when parsing an SDMX-ML document.
}
\examples{
#how to create a SDMXDotStatRequestBuilder
  requestBuilder <- SDMXDotStatRequestBuilder(
    regUrl = "http://www.myorg/registry",
    repoUrl = "http://www.myorg/repository")
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

