\name{SDMXGenericData-class}
\docType{class}
\alias{SDMXGenericData-class}
\alias{SDMXGenericData}
\alias{SDMXGenericData-method}


\title{Class "SDMXGenericData"}
\description{ A basic class to handle a SDMX-ML generic data set}
\section{Objects from the Class}{are never to be generated; will be used by the \link{readSDMX} main parser function}


\usage{
SDMXGenericData(xmlObj)
}

\arguments{
\item{xmlObj}{an object of class "XMLInternalDocument"}
}

\author{ Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{classes}