% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXGenericData.R,
%   R/SDMXGenericData-methods.R
\docType{class}
\name{SDMXGenericData}
\alias{SDMXGenericData}
\alias{SDMXGenericData-class}
\alias{SDMXGenericData,SDMXGenericData-method}
\title{Class "SDMXGenericData"}
\usage{
SDMXGenericData(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXGenericData"
}
\description{
A basic class to handle a SDMX-ML Generic data set
}
\section{Warning}{

This class is not useful in itself, but all SDMX non-abstract classes will 
encapsulate it as slot, when parsing an SDMX-ML document.
}

\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
