% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXItemScheme.R
\docType{class}
\name{SDMXItemScheme}
\alias{SDMXItemScheme}
\alias{SDMXItemScheme-class}
\title{Class "SDMXItemScheme"}
\description{
A basic abstract class to handle a SDMXItemScheme
}
\section{Warning}{

This class is not useful in itself, but all SDMX non-abstract "scheme" classes
should implement it. Added for the sake of complying with the SDMX information
structure model
}

\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
