\name{SDMXMessageGroup-class}
\docType{class}
\alias{SDMXMessageGroup-class}
\alias{SDMXMessageGroup}
\alias{SDMXMessageGroup-method}


\title{Class "SDMXMessageGroup"}
\description{ A basic class to handle a SDMX-ML message group}
\section{Objects from the Class}{are never to be generated; will be used by the \link{readSDMX} main parser function}

\usage{
SDMXMessageGroup(xmlObj)
}

\arguments{
\item{xmlObj}{an object of class "XMLInternalDocument"}
}

\author{ Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{classes}