% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXOrganisation.R,
%   R/SDMXOrganisation-methods.R
\docType{class}
\name{SDMXOrganisation}
\alias{SDMXOrganisation}
\alias{SDMXOrganisation-class}
\alias{SDMXOrganisation,SDMXOrganisation-method}
\title{Class "SDMXOrganisation"}
\usage{
SDMXOrganisation(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXOrganisation"
}
\description{
A basic class to handle a SDMX Concept
}
\section{Slots}{

\describe{
\item{\code{id}}{Object of class "character" giving the ID of the concept (required)}

\item{\code{uri}}{Object of class "character" giving the concept uri}

\item{\code{urn}}{Object of class "character" giving the concept urn}

\item{\code{Name}}{Object of class "list" giving the organisation name (by language) - required}

\item{\code{Description}}{Object of class "list" giving the organisation description (by language)}
}}

\section{Warning}{

This class is not useful in itself, but other classes such as \link{SDMXAgency}
will implement it.
}

\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
