% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-SDMXRequestParams.R, R/SDMXRequestParams-methods.R
\docType{class}
\name{SDMXRequestParams}
\alias{SDMXRequestParams}
\alias{SDMXRequestParams,SDMXRequestParams-method}
\alias{SDMXRequestParams-class}
\title{Class "SDMXRequestParams"}
\usage{
SDMXRequestParams(regUrl, repoUrl, agencyId, resource, resourceId, version,
                   flowRef, key, start, end, compliant)
}
\arguments{
\item{regUrl}{an object of class "character" giving the base Url of the SDMX service registry}

\item{repoUrl}{an object of class "character" giving the base Url of the SDMX service repository}

\item{agencyId}{an object of class "character" giving the agencyID}

\item{resource}{an object of class "character" giving the type of resource to be queried}

\item{resourceId}{an object of class "character" giving the resource to be queried}

\item{version}{an object of class "character" giving the resource version}

\item{flowRef}{an object of class "character" giving the flowRef to be queried}

\item{key}{an object of class "character" giving the key (SDMX url formatted) to be used for the query}

\item{start}{an object of class "character" giving the start time}

\item{end}{an object of class "character" giving the end time}

\item{compliant}{an object of class "logical" indicating if the web-service is compliant with the SDMX REST web-service specifications}
}
\description{
A class to handle a SDMX service request params
}
\section{Slots}{

\describe{
\item{\code{regUrl}}{an object of class "character" giving the base Url of the SDMX service registry}

\item{\code{repoUrl}}{an object of class "character" giving the base Url of the SDMX service repository}

\item{\code{agencyId}}{an object of class "character" giving the agencyID}

\item{\code{resource}}{an object of class "character" giving the type of resource to be queried}

\item{\code{resourceId}}{an object of class "character" giving the resource to be queried}

\item{\code{version}}{an object of class "character" giving the resource version}

\item{\code{flowRef}}{an object of class "character" giving the flowRef to be queried}

\item{\code{key}}{an object of class "character" giving the key (SDMX url formatted) to be used for the query}

\item{\code{start}}{an object of class "character" giving the start time}

\item{\code{end}}{an object of class "character" giving the end time}

\item{\code{compliant}}{an object of class "logical" indicating if the web-service is compliant with the SDMX REST web-service specifications}
}}
\section{Warning}{

This class is not useful in itself, but all SDMX non-abstract classes will
encapsulate it as slot, when parsing an SDMX-ML document.
}
\examples{
#how to create a SDMXRequestParams object
  params <- SDMXRequestParams(
   regUrl = "", repoUrl ="", agencyId ="", resource = "data", resourceId = "",
   version = "", flowRef = "", key = NULL, start = NULL, end = NULL, compliant = FALSE
  )
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

