\name{SDMXStructureSpecificData-class}
\docType{class}
\alias{SDMXStructureSpecificData-class}
\alias{SDMXStructureSpecificData}
\alias{SDMXStructureSpecificData-method}


\title{Class "SDMXStructureSpecificData"}
\description{ A basic class to handle a SDMX-ML structure specific data set}
\section{Objects from the Class}{are never to be generated; will be used by the \link{readSDMX} main parser function}


\usage{
SDMXStructureSpecificData(xmlObj)
}

\arguments{
\item{xmlObj}{an object of class "XMLInternalDocument"}
}

\author{ Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{classes}