% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXStructureSpecificData.R,
%   R/SDMXStructureSpecificData-methods.R
\docType{class}
\name{SDMXStructureSpecificData}
\alias{SDMXStructureSpecificData}
\alias{SDMXStructureSpecificData-class}
\alias{SDMXStructureSpecificData,SDMXStructureSpecificData-method}
\title{Class "SDMXStructureSpecificData"}
\usage{
SDMXStructureSpecificData(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXStructureSpecificData"
}
\description{
A basic class to handle a SDMX-ML StructureSpecificData data set
}
\section{Warning}{

This class is not useful in itself, but all SDMX non-abstract classes will 
encapsulate it as slot, when parsing an SDMX-ML document.
}

\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
