% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMXStructureType-methods.R
\docType{methods}
\name{getStructureType}
\alias{getStructureType}
\alias{getStructureType,SDMXStructureType-method}
\title{getStructureType}
\usage{
getStructureType(obj)

\S4method{getStructureType}{SDMXStructureType}(obj)
}
\arguments{
\item{obj}{object of class "SDMX"}
}
\description{
getStructureType
}
