% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-SDMXStructureType.R, R/SDMXStructureType-methods.R
\docType{class}
\name{SDMXStructureType}
\alias{SDMXStructureType}
\alias{SDMXStructureType,SDMXStructureType-method}
\alias{SDMXStructureType-class}
\title{Class "SDMXStructureType"}
\usage{
SDMXStructureType(xmlObj, resource)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}

\item{resource}{object of class "character" giving the REST resource to be
queried (required to distinguish between dataflows and datastructures in
SDMX 2.0)}
}
\value{
an object of class "SDMXStructureType"
}
\description{
A basic class to handle the type of a SDMX-ML Structure document
}
\section{Warning}{

This class is not useful in itself, but it will be used by \link{readSDMX} to
deal with SDMX-ML Structure documents.
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
\seealso{
\link{readSDMX}
}

