% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXTimeDimension.R,
%   R/SDMXTimeDimension-methods.R
\docType{class}
\name{SDMXTimeDimension}
\alias{SDMXTimeDimension}
\alias{SDMXTimeDimension-class}
\alias{SDMXTimeDimension,SDMXTimeDimension-method}
\title{Class "SDMXTimeDimension"}
\usage{
SDMXTimeDimension(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXTimeDimension"
}
\description{
A basic class to handle a SDMX TimeDimension
}
\section{Slots}{

\describe{
\item{\code{conceptRef}}{Object of class "character" giving the dimension conceptRef (required)}

\item{\code{conceptVersion}}{Object of class "character" giving the dimension concept version}

\item{\code{conceptAgency}}{Object of class "character" giving the dimension concept agency}

\item{\code{conceptSchemeRef}}{Object of class "character" giving the dimension conceptScheme ref}

\item{\code{conceptSchemeAgency}}{Object of class "character" giving the dimension conceptScheme agency}

\item{\code{codelist}}{Object of class "character" giving the codelist ref name}

\item{\code{codelistVersion}}{Object of class "character" giving the codelist ref version}

\item{\code{codelistAgency}}{Object of class "character" giving the codelist ref agency}

\item{\code{crossSectionalAttachDataset}}{Object of class "logical"}

\item{\code{crossSectionalAttachGroup}}{Object of class "logical"}

\item{\code{crossSectionalAttachSection}}{Object of class "logical"}

\item{\code{crossSectionalAttachObservation}}{Object of class "logical"}
}}

\section{Warning}{

This class is not useful in itself, but non-abstract classes willencapsulate 
it as slot, when parsing an SDMX-ML document (Concepts, or DataStructureDefinition)
}

\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
