% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXType.R, R/SDMXType-methods.R
\docType{class}
\name{SDMXType}
\alias{SDMXType}
\alias{SDMXType-class}
\alias{SDMXType,SDMXType-method}
\title{Class "SDMXType"}
\usage{
SDMXType(xmlObj)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}
}
\value{
an object of class "SDMXType"
}
\description{
A basic class to handle the type of a SDMX-ML document
}
\section{Slots}{

\describe{
\item{\code{type}}{Object of class "character" giving the type of the SDMX-ML document}
}}

\note{
At now, the following types have been implemented and successfully tested:
 - \code{StructureType},
 - \code{GenericDataType},
 - \code{CompactDataType},
 - \code{StructureSpecificDataType},
 - \code{CrossSectionalDataType},
 - \code{UtilityDataType},
 - \code{MessageGroupType}
}
\section{Warning}{

this class is not useful in itself, but all SDMX non-abstract classes will 
encapsulate it as slot, when parsing an SDMX-ML document.
}

\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
