\name{SDMXUtilityData-class}
\docType{class}
\alias{SDMXUtilityData-class}
\alias{SDMXUtilityData}
\alias{SDMXUtilityData-method}


\title{Class "SDMXUtilityData"}
\description{ A basic class to handle a SDMX-ML utility data set}
\section{Objects from the Class}{are never to be generated; will be used by the \link{readSDMX} main parser function}


\usage{
SDMXUtilityData(xmlObj)
}

\arguments{
\item{xmlObj}{an object of class "XMLInternalDocument"}
}

\author{ Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\keyword{classes}