% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SDMXData-methods.R
\docType{methods}
\name{setDSD}
\alias{setDSD}
\alias{setDSD,SDMXData-method}
\title{setDSD}
\usage{
set(obj, dsd)

\S4method{setDSD}{SDMXData}(obj, dsd)
}
\arguments{
\item{obj}{An object deriving from class "SDMXData"}

\item{dsd}{An object of class "SDMXDataStructureDefinition"}
}
\value{
the 'obj' object of class "SDMXData" enriched with the dsd
}
\description{
set the 'dsd' slot of a \code{SDMXData} object
}
\section{Methods (by class)}{
\itemize{
\item \code{SDMXData}: 
}}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
\seealso{
\link{SDMXData-class}
}

