\name{timedimension.SDMXComponents}
\Rdversion{1.1}
\alias{timedimension.SDMXComponents}
\title{timedimension.SDMXComponents}

\description{function used to parse the timedimension of a SDMX-ML document}

\usage{
  timedimension.SDMXComponents(xmlObj)
}

\arguments{
  \item{xmlObj}{an object of class "XMLInternalDocument"}
}

\author{Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}}

\section{Warning }{\code{timedimension.SDMXComponents} is a function used internally to parse the timedimension during instantiation of "SDMXComponents" objects (used in DataStructures / KeyFamilies)}

\seealso{\link{SDMXComponents-class} \link{SDMXDataStructure-class}}