% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMX-methods.R
\name{findNamespace}
\alias{findNamespace}
\title{findNamespace}
\usage{
findNamespace(namespaces, messageType)
}
\arguments{
\item{namespaces}{object of class \code{data.frame} giving the namespaces URIs
available in a SDMX-ML object, typically obtained with \link{getNamespaces}}

\item{messageType}{object of class \code{character} representing a message type}
}
\value{
an object of class "character" giving the namespace uri if found in the
        available namespaces
}
\description{
function used to find a specific namespace within the available 
             namespaces of an SDMX-ML object
}
\section{Warning}{

\code{findNamespace} is a function used internally as utility function in 
SDMX-ML object parsers.
}

\seealso{
\link{SDMX-class} \link{getNamespaces}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
