% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SDMX-methods.R
\name{getSoapRequestResult}
\alias{getSoapRequestResult}
\title{getSoapRequestResult}
\usage{
getSoapRequestResult(xmlObj)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}
}
\value{
an object of class "XMLInternalDocument derived from XML package
}
\description{
function used to extract the SDMX-ML message from a SOAP request
             response
}
\section{Warning}{

\code{getSoapRequestResult} is a function used internally by \link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
\seealso{
\link{SDMX-class} \link{readSDMX}
}

