% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SDMX-methods.R
\name{isSoapRequestEnvelope}
\alias{isSoapRequestEnvelope}
\title{isSoapRequestEnvelope}
\usage{
isSoapRequestEnvelope(xmlObj)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}
}
\value{
an object of class "logical"
}
\description{
function used to detect if the XML document corresponds to a SOAP
             request response
}
\section{Warning}{

\code{isSoapRequestEnvelope} is a function used internally by \link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
\seealso{
\link{SDMX-class} \link{readSDMX}
}

