\name{rsdmx}
\Rdversion{1.1}
\alias{rsdmx-package}
\alias{rsdmx}


\docType{package}

\title{A R package to read SDMX data and metadata}

\description{RSDMX is a package to parse/read SDMX documents in R. It provides classes and methods to read SDMX data and metadata.}

\details{
\tabular{ll}{
Package: \tab rsdmx\cr
Type: \tab Package\cr
Version
: \tab 0.4-1\cr
Date: \tab 2014-09-01\cr
License: \tab GPL(>=2.0)\cr
LazyLoad: \tab yes\cr
}

}

\author{Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}}


\keyword{SDMX}
\keyword{statistics}
\keyword{timeseries}
