% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdmx.R
\docType{package}
\name{rsdmx}
\alias{rsdmx}
\alias{rsdmx-package}
\title{Tools for Reading SDMX Data and Metadata}
\description{
RSDMX is a package to parse/read SDMX documents in R. It provides:
- a set of classes and methods to read data and metadata documents exchanged 
through theStatistical Data and Metadata Exchange (SDMX) framework. The package 
currently focuses on the SDMX XML standard format (SDMX-ML).
- an web-service interface to well-known SDMX data providers
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/opensdmx/rsdmx}
  \item \url{https://sdmx.org}
  \item Report bugs at \url{https://github.com/opensdmx/rsdmx/issues}
}

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
