% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdmx_logger.R
\docType{class}
\name{rsdmxLogger}
\alias{rsdmxLogger}
\title{rsdmxLogger}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a simple logger
}
\description{
rsdmxLogger
}
\note{
Logger class used internally by rsdmx
}
\section{Abstract Methods}{

\describe{
 \item{\code{INFO(text)}}{
   Logger to report information. Used internally
 }
 \item{\code{WARN(text)}}{
   Logger to report warnings. Used internally
 }
 \item{\code{ERROR(text)}}{
   Logger to report errors. Used internally
 }
}
}

\keyword{logger}
