% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSDMX.R
\name{saveSDMX}
\alias{saveSDMX}
\title{saveSDMX}
\usage{
saveSDMX(sdmxObj, file)
}
\arguments{
\item{sdmxObj}{an object of class \code{\link{SDMX-class}} to save}

\item{file}{a connection or the name of the file where the R object is saved to}
}
\description{
\code{saveSDMX} is the function to save R SDMX object
}
\examples{
            
 \donttest{
   sdmx <- system.file("extdata","Example_Eurostat_2.0.xml", package="rsdmx")
   sf <- tempfile(fileext = ".rda")
   saveSDMX(sdmx, sf)
 }
 
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
