% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMX-methods.R
\docType{methods}
\name{getNamespaces}
\alias{getNamespaces}
\alias{getNamespaces,SDMX-method}
\alias{getNamespaces,SDMX,ANY-method}
\title{getNamespaces}
\usage{
getNamespaces(obj)

\S4method{getNamespaces}{SDMX}(obj)
}
\arguments{
\item{obj}{An object deriving from class "SDMX"}
}
\value{
an object of class \code{data.frame} giving the id and uri for each 
        of the namespaces handled in the SDMX-ML document.
}
\description{
Access the namespaces of the SDMX-ML object
}
\seealso{
\link{SDMX-class}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
