% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXCompactData.R,
%   R/SDMXCompactData-methods.R
\docType{class}
\name{SDMXCompactData}
\alias{SDMXCompactData}
\alias{SDMXCompactData-class}
\alias{SDMXCompactData,SDMXCompactData-method}
\title{Class "SDMXCompactData"}
\usage{
SDMXCompactData(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXCompactData"
}
\description{
A basic class to handle a SDMX-ML compact data set
}
\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
