% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXComponents.R,
%   R/SDMXComponents-methods.R
\docType{class}
\name{SDMXComponents}
\alias{SDMXComponents}
\alias{SDMXComponents-class}
\alias{SDMXComponents,SDMXComponents-method}
\title{Class "SDMXComponents"}
\usage{
SDMXComponents(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXComponents"
}
\description{
A basic class to handle SDMX Components
}
\section{Slots}{

\describe{
\item{\code{Dimensions}}{Object of class "list" giving the list of dimensions (see \link{SDMXDimension})}

\item{\code{TimeDimension}}{Object of class "SDMXTimeDimension"}

\item{\code{PrimaryMeasure}}{Object of class "SDMXPrimaryMeasure"}

\item{\code{Attributes}}{Object of class "list" giving the list of attributes (see \link{SDMXAttribute})}
}}

\section{Warning}{

This class is not useful in itself, but all SDMX non-abstract classes will 
encapsulate it as slot, when parsing an SDMX-ML document (DataStructures, or 
DataStructureDefinitions)
}

\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
