% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXData.R, R/SDMXData-methods.R
\docType{class}
\name{SDMXData}
\alias{SDMXData}
\alias{SDMXData-class}
\alias{SDMXData,SDMXData-method}
\title{Class "SDMXData"}
\usage{
SDMXData(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXData"
}
\description{
An abstract class from which SDMX Data classes are derived
}
\section{Slots}{

\describe{
\item{\code{dsdRef}}{Object of class "character" giving the DSD Reference}

\item{\code{dsd}}{Object of class "SDMXDataStructureDefinition"}
}}

\section{Warning}{

This class is not useful in itself, but all SDMX Data classes in this package 
derive from it.                  
 
 @author Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}

\seealso{
\link{readSDMX}
}
