% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXMessageGroup.R,
%   R/SDMXMessageGroup-methods.R
\docType{class}
\name{SDMXMessageGroup}
\alias{SDMXMessageGroup}
\alias{SDMXMessageGroup-class}
\alias{SDMXMessageGroup,SDMXMessageGroup-method}
\title{Class "SDMXMessageGroup"}
\usage{
SDMXMessageGroup(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXMessageGroup"
}
\description{
A basic class to handle a SDMX-ML MessageGroup data set
}
\section{Warning}{

This class is not useful in itself, but all SDMX non-abstract classes will 
encapsulate it as slot, when parsing an SDMX-ML document.
}

\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
