% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SDMXPrimaryMeasure.R,
%   R/SDMXPrimaryMeasure-methods.R
\docType{class}
\name{SDMXPrimaryMeasure}
\alias{SDMXPrimaryMeasure}
\alias{SDMXPrimaryMeasure-class}
\alias{SDMXPrimaryMeasure,SDMXPrimaryMeasure-method}
\title{Class "SDMXPrimaryMeasure}
\usage{
SDMXPrimaryMeasure(xmlObj, namespaces)
}
\arguments{
\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}

\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
}
\value{
an object of class "SDMXPrimaryMeasure"
}
\description{
A basic class to handle a SDMX PrimaryMeasure
}
\section{Slots}{

\describe{
\item{\code{conceptRef}}{Object of class "character" giving the dimension conceptRef (required)}

\item{\code{conceptVersion}}{Object of class "character" giving the dimension concept version}

\item{\code{conceptAgency}}{Object of class "character" giving the dimension concept agency}

\item{\code{conceptSchemeRef}}{Object of class "character" giving the dimension conceptScheme ref}

\item{\code{conceptSchemeAgency}}{Object of class "character" giving the dimension conceptScheme agency}

\item{\code{codelist}}{Object of class "character" giving the codelist ref name}

\item{\code{codelistVersion}}{Object of class "character" giving the codelist ref version}

\item{\code{codelistAgency}}{Object of class "character" giving the codelist ref agency}
}}

\section{Warning}{

This class is not useful in itself, but all SDMX non-abstract classes will 
encapsulate it as slot, when parsing an SDMX-ML document.
}

\seealso{
\link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
