% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMXServiceProvider-methods.R
\name{setSDMXServiceProviders}
\alias{setSDMXServiceProviders}
\title{setSDMXServiceProviders}
\usage{
setSDMXServiceProviders()
}
\description{
function used internally by \pkg{rsdmx}, when loading the package, 
             to set the list of \link{SDMXServiceProvider} known by \pkg{rsdmx} 
             (hence known by \link{readSDMX} to query data/metadata in an easier 
             way). For internal use only (this function does not provide any 
             value for the end user, but it is here documented for transparency, 
             and to explain how the package works.)
}
\seealso{
\link{getSDMXServiceProviders} \link{addSDMXServiceProvider}
         \link{findSDMXServiceProvider} \link{readSDMX}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
